backend/requirements.txt
flask
openai
supabase
httpx
python-dotenv
ngrok